requirements.txt
cx_Oracle
python-dotenv
tabulate